requirements.txt
annotated-types==0.7.0
anyio==4.12.1
black==26.1.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
distro==1.9.0
google-auth==2.48.0
google-genai==1.63.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
isort==8.0.0
mypy_extensions==1.1.0
numpy==2.4.2
opencv-python==4.13.0.92
packaging==26.0
pathspec==1.0.4
pillow==12.1.1
platformdirs==4.9.2
pluggy==1.6.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytesseract==0.3.13
pytest==9.0.2
python-dotenv==1.2.1
pytokens==0.4.1
pywin32==311
requests==2.32.5
rsa==4.9.1
ruff==0.15.2
sniffio==1.3.1
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==15.0.1
